refactor(eslint-plugin): move angular-rule utils to a separate directory
