+++ packages/eslint-plugin/utils/angular-rule/false-values.js
